requirements.txt
absl-py==0.8.0
boto==2.49.0
boto3==1.12.39
botocore==1.15.39
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
cycler==0.10.0
docutils==0.15.2
gensim==3.8.2
grpcio==1.23.0
idna==2.9
jmespath==0.9.5
joblib==0.14.1
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
nltk==3.5
numpy==1.17.2
opencv-python==4.2.0.34
Pillow==6.1.0
protobuf==3.9.1
pyenchant==3.0.1
pyparsing==2.4.2
python-dateutil==2.8.0
regex==2020.4.4
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.12.0
smart-open==1.11.1
tensorboard==2.0.0
tensorboardX==1.8
torch==1.0.0
torchvision==0.2.1
tqdm==4.45.0
urllib3==1.25.8
Werkzeug==0.16.0
